requirements.txt
azure-functions==1.19.0
bcrypt==4.3.0
mysql-connector-python==8.3.0
requests==2.32.3
pyjwt==2.8.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.0